corrected mistakes
