+++ .gitignore
@@ -2,4 +2,9 @@
 .idea
 out.txt
 1.txt
+test.txt
+test1.txt
+test2.txt
+coverage.out
 bin/
+tests/
