+++ go.mod
@@ -1,3 +1,5 @@
-module example
+module main
 
 go 1.19
+
+require github.com/google/uuid v1.3.0 // indirect
